Rework testcase modules for meet functions
